+++ run_instruments.py
@@ -10,25 +10,38 @@ from structs import Params
 
 log = logging.getLogger(__name__)
 
-def run_single(pm, laser, params, dryrun=False):
-    log.info("--- Start a test ---")
+power_limit = {"10": 10e-3,
+             "0"  : 1.9999e-3,
+             "-10": 199.99e-6,
+             "-20": 19.999e-6,
+             "-30": 1.9999e-6,
+             "-40": 199.99e-9,
+             "-50": 19.999e-9,
+             "-60": 1.9999e-9,
+             "-70": 199.99e-12}
+
+def run_sweep(pm, laser, params, dryrun=False):
+    log.info(f"Type: {params.sweep_type}")
+    log.info("--- Running instruments ---")
     
     tls_wl_start = params.wl_start - params.padding
-    tsl_wl_stop  = params.wl_stop  + params.padding
+    tls_wl_stop  = params.wl_stop  + params.padding
 
     # ----- Power Meter -----
     pm.write(":INIT1:CONT 0")
     pm.write(":SENS1:FUNC:STAT LOGG, STOP")
-    pm.write(f":SENSE1:CHAN1:POW:WAVE {tsl_wl_stop:.3f} NM")
+    pm.write(f":SENSE1:CHAN1:POW:WAVE {tls_wl_stop:.3f} NM")
     pm.write(f":SENSE1:CHAN1:POW:ATIME {params.at_us} US")
     pm.write(":SENSE1:CHAN1:CORR 0")
     pm.write(":SENSE1:CHAN1:POW:RANGE:AUTO  0")
-    pm.write(":SENSE1:CHAN1:POW:RANGE  10 DBM")
+    pm.write(f":SENSE1:CHAN1:POW:RANGE  {params.pm_range} DBM")
     pm.write(":SENSE1:CHAN1:POW:UNIT  1")   # W (faster)
     pm.write(":TRIG1:CHAN1:OUTP DIS")
     pm.write(":TRIG1:CHAN1:INP  CME")
     pm.write(":TRIG:CONF PASS")
     pm.write(f":SENSE1:CHAN1:FUNC:PAR:LOGG {params.num_data}, {params.at_us} US")
+
+    # log.info(f"[PM] Power Range: {pm.query(":SENSE1:CHAN1:POW:RANGE?")}")
 
     # ----- Laser -----
         
@@ -44,7 +57,7 @@ def run_single(pm, laser, params, dryrun=False):
     laser.write(":SOURCE0:WAV:SWE:REP ONEW")
     laser.write(f":SOURCE0:WAV:SWE:SPE      {params.speed} NM/S")
     laser.write(f":SOURCE0:WAV:SWE:STAR     {tls_wl_start:.3f} NM")
-    laser.write(f":SOURCE0:WAV:SWE:STOP     {tsl_wl_stop:.3f} NM")
+    laser.write(f":SOURCE0:WAV:SWE:STOP     {tls_wl_stop:.3f} NM")
 
     laser_check_param = (laser.query(":SOUR0:WAV:SWE:CHEC?")).split(',')
     if int(laser_check_param[0]) != 0:
@@ -55,7 +68,7 @@ def run_single(pm, laser, params, dryrun=False):
     if (n := int(laser.query(":SYST:ERR:COUN?"))) > 0:
         for _ in range(n):
             log.error(f"[LASER] System error: {laser.query(':SYST:ERR?')}")
-        return None
+            sys.exit(1)
     
     # TODO: check behavior of power meter. Check if ":SYST:ERR:COUN?" works
     # pm_check_error1 = (pm.query(":SYST:ERR?")).split(',')
@@ -95,9 +108,17 @@ def run_single(pm, laser, params, dryrun=False):
     log.info(f"[LASER] Check if any error occurred: {laser.query(':SYST:ERR?')}")
     log.info(f"[PM] Check if any error occurred: {pm.query(':SYST:ERR?')}")
 
-    log.info("--- Test finished ---\n")
+    log.info("--- End of running instruments ---\n")
 
-    arr_dbm  = 10 * np.log10(arr_w / 1e-3)  # convert: W -> dBm
+    upper_limit = power_limit[str(params.pm_range)]
+    arr_w[arr_w > upper_limit] = np.nan
+    
+    if np.all(np.isnan(arr_w)):
+        log.error("All measurements are overflown.")
+    elif np.any(np.isnan(arr_w)):
+        log.warning("There are some overflown measurements.")
+        
+    arr_dbm  = 10 * np.log10(arr_w.astype(np.float64) / 1e-3)  # convert: W -> dBm
     arr_dbm *= -1                           # power loss 
     
     # Generate x-axis
@@ -106,10 +127,8 @@ def run_single(pm, laser, params, dryrun=False):
 
     i_lo = np.searchsorted(wav_range, params.wl_start - params.step_pm * 1e-3, side='left')
     i_hi = np.searchsorted(wav_range, params.wl_stop  + params.step_pm * 1e-3, side='right')
-
-    data = np.column_stack((wav_range[i_lo:i_hi], arr_dbm[i_lo:i_hi]))
     
-    return data
+    return (wav_range[i_lo:i_hi], arr_dbm[i_lo:i_hi])
 
 if __name__ == "__main__":
     pm, laser = prep_inst()